Move incident service into core namespace
- Get options from config
