+++ src/FunderMaps.Core/IncidentReport/IncidentOptions.cs
@@ -1,0 +1,25 @@
+using System.Collections.Generic;
+
+namespace FunderMaps.Core.IncidentReport
+{
+    /// <summary>
+    ///     Incident options.
+    /// </summary>
+    public sealed record IncidentOptions
+    {
+        /// <summary>
+        ///     Configuration section key.
+        /// </summary>
+        public const string Section = "Incident";
+
+        /// <summary>
+        ///     Incident client identifier.
+        /// </summary>
+        public int ClientId { get; set; }
+
+        /// <summary>
+        ///     The notification recipients.
+        /// </summary>
+        public List<string> Recipients { get; set; }
+    }
+}
