+++ src/FunderMaps.Core/IncidentReport/IncidentService.cs
@@ -2,45 +2,45 @@ using FunderMaps.Core.Entities;
 using FunderMaps.Core.Interfaces;
 using FunderMaps.Core.Interfaces.Repositories;
 using FunderMaps.Core.Notification;
-using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Options;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace FunderMaps.Core.Services
+namespace FunderMaps.Core.IncidentReport
 {
     /// <summary>
     ///     Service to the incidents.
     /// </summary>
     internal class IncidentService : IIncidentService
     {
+        private readonly IncidentOptions _options;
         private readonly Core.AppContext _appContext;
-        private readonly IConfiguration _configuration;
         private readonly IContactRepository _contactRepository;
         private readonly IIncidentRepository _incidentRepository;
         private readonly IGeocoderTranslation _geocoderTranslation;
         private readonly INotifyService _notifyService;
 
-        // FUTURE: IOptions
         /// <summary>
         ///     Create new instance.
         /// </summary>
         public IncidentService(
+            IOptions<IncidentOptions> options,
             Core.AppContext appContext,
-            IConfiguration configuration,
             IContactRepository contactRepository,
             IIncidentRepository incidentRepository,
             IGeocoderTranslation geocoderTranslation,
             INotifyService notifyService)
         {
+            _options = options?.Value ?? throw new ArgumentNullException(nameof(options));
             _appContext = appContext ?? throw new ArgumentNullException(nameof(appContext));
-            _configuration = configuration ?? throw new ArgumentNullException(nameof(configuration));
             _contactRepository = contactRepository ?? throw new ArgumentNullException(nameof(contactRepository));
             _incidentRepository = incidentRepository ?? throw new ArgumentNullException(nameof(incidentRepository));
             _geocoderTranslation = geocoderTranslation ?? throw new ArgumentNullException(nameof(geocoderTranslation));
             _notifyService = notifyService ?? throw new ArgumentNullException(nameof(notifyService));
         }
 
+        // FUTURE: split logic, hard to read.
         /// <summary>
         ///     Register a new incident.
         /// </summary>
@@ -51,7 +51,7 @@ namespace FunderMaps.Core.Services
             Address address = await _geocoderTranslation.GetAddressIdAsync(incident.Address);
 
             incident.Address = address.Id;
-            incident.ClientId = int.Parse(_configuration["Incident:ClientId"]);
+            incident.ClientId = _options.ClientId;
             incident.Meta = meta;
 
             await _contactRepository.AddAsync(incident.ContactNavigation);
@@ -61,7 +61,7 @@ namespace FunderMaps.Core.Services
 
             await _notifyService.DispatchNotifyAsync("incident_notify", new()
             {
-                Recipients = new List<string> { "info@fundermaps.com", "info@kcaf.nl" }, // TODO: Retrieve from config
+                Recipients = _options.Recipients,
                 Items = new Dictionary<string, string> { { "id", id } },
             });
